refact : clean company domain
