+++ database/seeders/CompanySeeder.php
@@ -31,5 +31,14 @@ class CompanySeeder extends Seeder
             'created_at' => date('Y-m-d H:i:s'),
             'is_deleted' => 0
         ]);
+
+        DB::table('companies_associations')->insert([
+            'name' => 'Kanvas',
+            'users_id' => 1,
+            'system_modules_id' => 1,
+            'language' => 'EN',
+            'created_at' => date('Y-m-d H:i:s'),
+            'is_deleted' => 0
+        ]);
     }
 }
