+++ src/Kanvas/Companies/Models/Companies.php
@@ -9,8 +9,7 @@ use Illuminate\Database\Eloquent\Relations\HasOne;
 use Kanvas\Apps\Models\Apps;
 use Kanvas\Companies\Enums\Defaults;
 use Kanvas\Companies\Factories\CompaniesFactory;
-use Kanvas\CompanyGroup\Branches\Models\CompaniesBranches;
-use Kanvas\CompanyGroup\Groups\Models\CompaniesGroups;
+use Kanvas\CompaniesGroups\Models\CompaniesGroups;
 use Kanvas\Enums\StateEnums;
 use Kanvas\Models\BaseModel;
 use Kanvas\SystemModules\Models\SystemModules;
