+++ src/Kanvas/Companies/Observers/CompaniesObserver.php
@@ -6,9 +6,9 @@ namespace Kanvas\Companies\Observers;
 use Illuminate\Support\Str;
 use Kanvas\Apps\Models\Apps;
 use Kanvas\Companies\Models\Companies;
-use Kanvas\CompanyGroup\Branches\Actions\CreateCompanyBranchActions;
-use Kanvas\CompanyGroup\Branches\DataTransferObject\CompaniesBranchPostData;
-use Kanvas\CompanyGroup\Groups\Actions\CreateCompanyGroupActions;
+use Kanvas\Companies\Branches\Actions\CreateCompanyBranchActions;
+use Kanvas\Companies\Branches\DataTransferObject\CompaniesBranchPostData;
+use Kanvas\Companies\Groups\Actions\CreateCompanyGroupActions;
 use Kanvas\Enums\AppEnums;
 use Kanvas\Enums\StateEnums;
 use Kanvas\Users\Actions\AssignRole;
